[FEATURE] Adds Query Include/Exclude Package
Change Details
-------------
A project suggested it would be great to have the ability to specify
which posts are included or excluded from a query-loop block. This
package aims to provide this feature. The functionality already exists
in core, but there's no interface to control it. At the moment this
package provides a simple interface.

Further enhancements could be acheievd in the future, such as limiting
the number of selectable posts to include to the same limit as the query
and only returning posts that meet other filters selected in the query
for the include/exclude.

Also, there's potential performance issues if the number of posts or pages
is excessive so we could look at more lazily loading these.