[FIX] Template Persistance Disabling
The filter being used to disable filtering was being called very early in the lifecycle. This made it near impossible for a theme to actually utilise this filter. Moving it inside the save-post hook allows the filter to be utilised.

Also added a default value so that template persistance is only effective on local environments, with the filter being used to ovveride this default value.